chore(test): move tests to the source dir
