+++ onnxscript/tests/README.md
@@ -1,0 +1,4 @@
+# tests
+
+This directory contains integration tests and common test utilities (`common/`).
+Module unit tests should be creating next to the module source with the name `<module name>_test.py`.
