+++ onnxscript/tests/functions/gemmgelu_test.py
@@ -7,8 +7,8 @@ import unittest
 
 import numpy as np
 
-from onnxscript.test.common import onnx_script_test_case
-from onnxscript.test.functions import gemmgelu
+from onnxscript.tests.common import onnx_script_test_case
+from onnxscript.tests.functions import gemmgelu
 
 
 class TestGemmGelu(onnx_script_test_case.OnnxScriptTestCase):
