+++ onnxscript/tests/functions/if_test.py
@@ -7,8 +7,8 @@ import unittest
 
 import numpy as np
 
-from onnxscript.test.common import onnx_script_test_case
-from onnxscript.test.models import if_statement
+from onnxscript.tests.common import onnx_script_test_case
+from onnxscript.tests.models import if_statement
 
 
 class TestOnnxIf(onnx_script_test_case.OnnxScriptTestCase):
