+++ onnxscript/tests/models/eg1.py
