+++ onnxscript/tests/onnx_backend_test.py
@@ -22,7 +22,7 @@ import onnxscript
 from onnxscript import evaluator
 from onnxscript.backend.onnx_backend import enumerate_onnx_tests
 from onnxscript.backend.onnx_export import export2python
-from onnxscript.test.models import type_double
+from onnxscript.tests.models import type_double
 
 
 def print_code(code, begin=1):
