+++ onnxscript/type_annotation_test.py
@@ -11,7 +11,7 @@ from packaging.version import Version
 from onnxscript import script
 from onnxscript.onnx_opset import opset15 as op
 from onnxscript.onnx_types import FLOAT
-from onnxscript.test.common import testutils
+from onnxscript.tests.common import testutils
 
 
 class TypeAnnotationTester(testutils.TestBase):
